requirements.txt
Django
djangorestframework
python-dotenv
Pillow